requirements.txt
altgraph==0.17.4
apturl==0.5.2
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
blinker==1.9.0
boto3==1.36.6
botocore==1.36.6
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-PyMongo==2.3.0
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jmespath==1.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.2
oauthlib==3.2.0
olefile==0.46
packaging==24.2
pandas==2.2.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.5.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.10.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.4
pytest-flask==1.3.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
s3transfer==0.11.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
ssh-import-id==5.11
tomli==2.2.1
Twisted==22.1.0
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unotools==0.3.3
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
